Split up HtmlElement tests
